+++ stroom-pipeline/src/main/java/stroom/pipeline/refdata/ReferenceDataResourceImpl.java
@@ -1,6 +1,8 @@
 package stroom.pipeline.refdata;
 
 import stroom.pipeline.refdata.store.RefStoreEntry;
+import stroom.util.logging.LogUtil;
+import stroom.util.time.StroomDuration;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -30,5 +32,22 @@ public class ReferenceDataResourceImpl implements ReferenceDataResource {
     public String lookup(final RefDataLookupRequest refDataLookupRequest) {
         return referenceDataServiceProvider.get()
                 .lookup(refDataLookupRequest);
+    }
+
+    @Override
+    public void purge(final String purgeAge) {
+        StroomDuration purgeAgeDuration;
+        try {
+            purgeAgeDuration = StroomDuration.parse(purgeAge);
+        } catch (Exception e) {
+            throw new IllegalArgumentException(LogUtil.message(
+                    "Can't parse purgeAge [{}]", purgeAge), e);
+        }
+        try {
+            referenceDataServiceProvider.get()
+                    .purge(purgeAgeDuration);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
     }
 }